requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pandas==2.1.3
numpy==1.26.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
aiohttp==3.9.1
